Add a cast to quiet a compiler warning.

+++ plugins/sudoers/iolog.c
@@ -286,7 +286,7 @@ sudoers_io_open(unsigned int version, sudo_conv_t conversation,
     gettimeofday(&last_time, NULL);
 
     /* XXX - log more stuff?  window size? environment? */
-    fprintf(io_logfile, "%ld:%s:%s:%s:%s\n", last_time.tv_sec, user_name,
+    fprintf(io_logfile, "%ld:%s:%s:%s:%s\n", (long)last_time.tv_sec, user_name,
         runas_pw->pw_name, runas_gr ? runas_gr->gr_name : "", user_tty);
     fprintf(io_logfile, "%s\n", user_cwd);
     fprintf(io_logfile, "%s%s%s\n", user_cmnd, user_args ? " " : "",
